requirements.txt
openai==1.12.0
flask==3.0.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
flask-cors==4.0.0